refactor: split reactutil.go into per-topic files
The 3655-line reactutil.go was hard to navigate and review. Split into 15
focused files (text/naming/glob/ast/pragma/pragma_binding/component_*/jsx*/
resolve/create_element/components_map/prop_wrappers) without modifying any
declaration. All exports preserved; the 60+ importers do not change.
